feat(core): Expose bundler plugin primitives via `createSentryBuildPluginManager` (#713)
