+++ packages/bundler-plugin-core/src/index.ts
@@ -1,16 +1,18 @@
-import SentryCli from "@sentry/cli";
 import { transformAsync } from "@babel/core";
 import componentNameAnnotatePlugin from "@sentry/babel-plugin-component-annotate";
+import SentryCli from "@sentry/cli";
+import { logger } from "@sentry/utils";
 import * as fs from "fs";
-import * as path from "path";
+import { glob } from "glob";
 import MagicString from "magic-string";
+import * as path from "path";
 import { createUnplugin, TransformResult, UnpluginOptions } from "unplugin";
-import { normalizeUserOptions, validateOptions } from "./options-mapping";
+import { createSentryBuildPluginManager } from "./api-primitives";
 import { createDebugIdUploadFunction } from "./debug-id-upload";
 import { releaseManagementPlugin } from "./plugins/release-management";
+import { fileDeletionPlugin } from "./plugins/sourcemap-deletion";
 import { telemetryPlugin } from "./plugins/telemetry";
-import { createLogger, Logger } from "./sentry/logger";
-import { allowedToSendTelemetry, createSentryInstance } from "./sentry/telemetry";
+import { Logger } from "./sentry/logger";
 import { Options, SentrySDKBuildFlags } from "./types";
 import {
   generateGlobalInjectorCode,
@@ -22,11 +24,6 @@ import {
   stringToUUID,
   stripQueryAndHashFromPath,
 } from "./utils";
-import * as dotenv from "dotenv";
-import { glob } from "glob";
-import { logger } from "@sentry/utils";
-import { fileDeletionPlugin } from "./plugins/sourcemap-deletion";
-import { closeSession, DEFAULT_ENVIRONMENT, makeSession } from "@sentry/core";
 
 interface SentryUnpluginFactoryOptions {
   releaseInjectionPlugin: (injectionCode: string) => UnpluginOptions;
@@ -36,6 +33,7 @@ interface SentryUnpluginFactoryOptions {
   debugIdUploadPlugin: (
     upload: (buildArtifacts: string[]) => Promise<void>,
     logger: Logger,
+    createDependencyOnBuildArtifacts: () => () => void,
     webpack_forceExitOnBuildComplete?: boolean
   ) => UnpluginOptions;
   bundleSizeOptimizationsPlugin: (buildFlags: SentrySDKBuildFlags) => UnpluginOptions;
@@ -76,40 +74,18 @@ export function sentryUnpluginFactory({
   bundleSizeOptimizationsPlugin,
 }: SentryUnpluginFactoryOptions) {
   return createUnplugin<Options | undefined, true>((userOptions = {}, unpluginMetaContext) => {
-    const logger = createLogger({
-      prefix:
+    const sentryBuildPluginManager = createSentryBuildPluginManager(userOptions, {
+      loggerPrefix:
         userOptions._metaOptions?.loggerPrefixOverride ??
         `[sentry-${unpluginMetaContext.framework}-plugin]`,
-      silent: userOptions.silent ?? false,
-      debug: userOptions.debug ?? false,
+      buildTool: unpluginMetaContext.framework,
     });
 
-    // Not a bulletproof check but should be good enough to at least sometimes determine
-    // if the plugin is called in dev/watch mode or for a  prod build. The important part
-    // here is to avoid a false positive. False negatives are okay.
-    const isDevMode = process.env["NODE_ENV"] === "development";
-
-    try {
-      const dotenvFile = fs.readFileSync(
-        path.join(process.cwd(), ".env.sentry-build-plugin"),
-        "utf-8"
-      );
-      // NOTE: Do not use the dotenv.config API directly to read the dotenv file! For some ungodly reason, it falls back to reading `${process.cwd()}/.env` which is absolutely not what we want.
-      const dotenvResult = dotenv.parse(dotenvFile);
-
-      // Vite has a bug/behaviour where spreading into process.env will cause it to crash
-      // https://github.com/vitest-dev/vitest/issues/1870#issuecomment-1501140251
-      Object.assign(process.env, dotenvResult);
-
-      logger.info('Using environment variables configured in ".env.sentry-build-plugin".');
-    } catch (e: unknown) {
-      // Ignore "file not found" errors but throw all others
-      if (typeof e === "object" && e && "code" in e && e.code !== "ENOENT") {
-        throw e;
-      }
-    }
-
-    const options = normalizeUserOptions(userOptions);
+    const {
+      logger,
+      normalizedOptions: options,
+      bundleSizeOptimizationReplacementValues,
+    } = sentryBuildPluginManager;
 
     if (options.disable) {
       return [
@@ -117,87 +93,6 @@ export function sentryUnpluginFactory({
           name: "sentry-noop-plugin",
         },
       ];
-    }
-
-    const shouldSendTelemetry = allowedToSendTelemetry(options);
-    const { sentryScope, sentryClient } = createSentryInstance(
-      options,
-      shouldSendTelemetry,
-      unpluginMetaContext.framework
-    );
-
-    const { release, environment = DEFAULT_ENVIRONMENT } = sentryClient.getOptions();
-
-    const sentrySession = makeSession({ release, environment });
-    sentryScope.setSession(sentrySession);
-    // Send the start of the session
-    sentryClient.captureSession(sentrySession);
-
-    let sessionHasEnded = false; // Just to prevent infinite loops with beforeExit, which is called whenever the event loop empties out
-
-    function endSession() {
-      if (sessionHasEnded) {
-        return;
-      }
-
-      closeSession(sentrySession);
-      sentryClient.captureSession(sentrySession);
-      sessionHasEnded = true;
-    }
-
-    // We also need to manually end sessions on errors because beforeExit is not called on crashes
-    process.on("beforeExit", () => {
-      endSession();
-    });
-
-    // Set the User-Agent that Sentry CLI will use when interacting with Sentry
-    process.env[
-      "SENTRY_PIPELINE"
-    ] = `${unpluginMetaContext.framework}-plugin/${__PACKAGE_VERSION__}`;
-
-    /**
-     * Handles errors caught and emitted in various areas of the plugin.
-     *
-     * Also sets the sentry session status according to the error handling.
-     *
-     * If users specify their custom `errorHandler` we'll leave the decision to throw
-     * or continue up to them. By default, @param throwByDefault controls if the plugin
-     * should throw an error (which causes a build fail in most bundlers) or continue.
-     */
-    function handleRecoverableError(unknownError: unknown, throwByDefault: boolean) {
-      sentrySession.status = "abnormal";
-      try {
-        if (options.errorHandler) {
-          try {
-            if (unknownError instanceof Error) {
-              options.errorHandler(unknownError);
-            } else {
-              options.errorHandler(new Error("An unknown error occured"));
-            }
-          } catch (e) {
-            sentrySession.status = "crashed";
-            throw e;
-          }
-        } else {
-          // setting the session to "crashed" b/c from a plugin perspective this run failed.
-          // However, we're intentionally not rethrowing the error to avoid breaking the user build.
-          sentrySession.status = "crashed";
-          if (throwByDefault) {
-            throw unknownError;
-          }
-          logger.error("An error occurred. Couldn't finish all operations:", unknownError);
-        }
-      } finally {
-        endSession();
-      }
-    }
-
-    if (!validateOptions(options, logger)) {
-      // Throwing by default to avoid a misconfigured plugin going unnoticed.
-      handleRecoverableError(
-        new Error("Options were not set correctly. See output above for more details."),
-        true
-      );
     }
 
     if (process.cwd().match(/\\node_modules\\|\/node_modules\//)) {
@@ -210,84 +105,12 @@ export function sentryUnpluginFactory({
 
     plugins.push(
       telemetryPlugin({
-        sentryClient,
-        sentryScope,
-        logger,
-        shouldSendTelemetry,
+        sentryBuildPluginManager,
       })
     );
 
-    // We have multiple plugins depending on generated source map files. (debug ID upload, legacy upload)
-    // Additionally, we also want to have the functionality to delete files after uploading sourcemaps.
-    // All of these plugins and the delete functionality need to run in the same hook (`writeBundle`).
-    // Since the plugins among themselves are not aware of when they run and finish, we need a system to
-    // track their dependencies on the generated files, so that we can initiate the file deletion only after
-    // nothing depends on the files anymore.
-    const dependenciesOnSourcemapFiles = new Set<symbol>();
-    const sourcemapFileDependencySubscribers: (() => void)[] = [];
-
-    function notifySourcemapFileDependencySubscribers() {
-      sourcemapFileDependencySubscribers.forEach((subscriber) => {
-        subscriber();
-      });
-    }
-
-    function createDependencyOnSourcemapFiles() {
-      const dependencyIdentifier = Symbol();
-      dependenciesOnSourcemapFiles.add(dependencyIdentifier);
-
-      return function freeDependencyOnSourcemapFiles() {
-        dependenciesOnSourcemapFiles.delete(dependencyIdentifier);
-        notifySourcemapFileDependencySubscribers();
-      };
-    }
-
-    /**
-     * Returns a Promise that resolves when all the currently active dependencies are freed again.
-     *
-     * It is very important that this function is called as late as possible before wanting to await the Promise to give
-     * the dependency producers as much time as possible to register themselves.
-     */
-    function waitUntilSourcemapFileDependenciesAreFreed() {
-      return new Promise<void>((resolve) => {
-        sourcemapFileDependencySubscribers.push(() => {
-          if (dependenciesOnSourcemapFiles.size === 0) {
-            resolve();
-          }
-        });
-
-        if (dependenciesOnSourcemapFiles.size === 0) {
-          resolve();
-        }
-      });
-    }
-
-    if (options.bundleSizeOptimizations) {
-      const { bundleSizeOptimizations } = options;
-      const replacementValues: SentrySDKBuildFlags = {};
-
-      if (bundleSizeOptimizations.excludeDebugStatements) {
-        replacementValues["__SENTRY_DEBUG__"] = false;
-      }
-      if (bundleSizeOptimizations.excludeTracing) {
-        replacementValues["__SENTRY_TRACING__"] = false;
-      }
-      if (bundleSizeOptimizations.excludeReplayCanvas) {
-        replacementValues["__RRWEB_EXCLUDE_CANVAS__"] = true;
-      }
-      if (bundleSizeOptimizations.excludeReplayIframe) {
-        replacementValues["__RRWEB_EXCLUDE_IFRAME__"] = true;
-      }
-      if (bundleSizeOptimizations.excludeReplayShadowDom) {
-        replacementValues["__RRWEB_EXCLUDE_SHADOW_DOM__"] = true;
-      }
-      if (bundleSizeOptimizations.excludeReplayWorker) {
-        replacementValues["__SENTRY_EXCLUDE_REPLAY_WORKER__"] = true;
-      }
-
-      if (Object.keys(replacementValues).length > 0) {
-        plugins.push(bundleSizeOptimizationsPlugin(replacementValues));
-      }
+    if (Object.keys(bundleSizeOptimizationReplacementValues).length > 0) {
+      plugins.push(bundleSizeOptimizationsPlugin(bundleSizeOptimizationReplacementValues));
     }
 
     if (!options.release.inject) {
@@ -306,152 +129,39 @@ export function sentryUnpluginFactory({
       plugins.push(releaseInjectionPlugin(injectionCode));
     }
 
-    if (options.moduleMetadata || options.applicationKey) {
-      let metadata: Record<string, unknown> = {};
-
-      if (options.applicationKey) {
-        // We use different keys so that if user-code receives multiple bundling passes, we will store the application keys of all the passes.
-        // It is a bit unfortunate that we have to inject the metadata snippet at the top, because after multiple
-        // injections, the first injection will always "win" because it comes last in the code. We would generally be
-        // fine with making the last bundling pass win. But because it cannot win, we have to use a workaround of storing
-        // the app keys in different object keys.
-        // We can simply use the `_sentryBundlerPluginAppKey:` to filter for app keys in the SDK.
-        metadata[`_sentryBundlerPluginAppKey:${options.applicationKey}`] = true;
-      }
-
-      if (typeof options.moduleMetadata === "function") {
-        const args = {
-          org: options.org,
-          project: options.project,
-          release: options.release.name,
-        };
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        metadata = { ...metadata, ...options.moduleMetadata(args) };
-      } else {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-        metadata = { ...metadata, ...options.moduleMetadata };
-      }
-
-      const injectionCode = generateModuleMetadataInjectorCode(metadata);
+    if (Object.keys(sentryBuildPluginManager.bundleMetadata).length > 0) {
+      const injectionCode = generateModuleMetadataInjectorCode(
+        sentryBuildPluginManager.bundleMetadata
+      );
       plugins.push(moduleMetadataInjectionPlugin(injectionCode));
     }
-    // https://turbo.build/repo/docs/reference/system-environment-variables#environment-variables-in-tasks
-    const isRunningInTurborepo = Boolean(process.env["TURBO_HASH"]);
-    const getTurborepoEnvPassthroughWarning = (envVarName: string) =>
-      isRunningInTurborepo
-        ? `\nYou seem to be using Turborepo, did you forget to put ${envVarName} in \`passThroughEnv\`? https://turbo.build/repo/docs/reference/configuration#passthroughenv`
-        : "";
-    if (!options.release.name) {
-      logger.debug(
-        "No release name provided. Will not create release. Please set the `release.name` option to identify your release."
-      );
-    } else if (isDevMode) {
-      logger.debug("Running in development mode. Will not create release.");
-    } else if (!options.authToken) {
-      logger.warn(
-        "No auth token provided. Will not create release. Please set the `authToken` option. You can find information on how to generate a Sentry auth token here: https://docs.sentry.io/api/auth/" +
-          getTurborepoEnvPassthroughWarning("SENTRY_AUTH_TOKEN")
-      );
-    } else if (!options.org && !options.authToken.startsWith("sntrys_")) {
-      logger.warn(
-        "No organization slug provided. Will not create release. Please set the `org` option to your Sentry organization slug." +
-          getTurborepoEnvPassthroughWarning("SENTRY_ORG")
-      );
-    } else if (!options.project) {
-      logger.warn(
-        "No project provided. Will not create release. Please set the `project` option to your Sentry project slug." +
-          getTurborepoEnvPassthroughWarning("SENTRY_PROJECT")
-      );
-    } else {
-      plugins.push(
-        releaseManagementPlugin({
-          logger,
-          releaseName: options.release.name,
-          shouldCreateRelease: options.release.create,
-          shouldFinalizeRelease: options.release.finalize,
-          include: options.release.uploadLegacySourcemaps,
-          // setCommits has a default defined by the options mappings
-          // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-          setCommitsOption: options.release.setCommits!,
-          deployOptions: options.release.deploy,
-          dist: options.release.dist,
-          handleRecoverableError: handleRecoverableError,
-          sentryScope,
-          sentryClient,
-          sentryCliOptions: {
-            authToken: options.authToken,
-            org: options.org,
-            project: options.project,
-            silent: options.silent,
-            url: options.url,
-            vcsRemote: options.release.vcsRemote,
-            headers: options.headers,
-          },
-          createDependencyOnSourcemapFiles,
-        })
-      );
-    }
+
+    plugins.push(
+      releaseManagementPlugin({
+        sentryBuildPluginManager,
+      })
+    );
 
     if (!options.sourcemaps?.disable) {
       plugins.push(debugIdInjectionPlugin(logger));
     }
 
-    if (options.sourcemaps?.disable) {
-      logger.debug(
-        "Source map upload was disabled. Will not upload sourcemaps using debug ID process."
-      );
-    } else if (isDevMode) {
-      logger.debug("Running in development mode. Will not upload sourcemaps.");
-    } else if (!options.authToken) {
-      logger.warn(
-        "No auth token provided. Will not upload source maps. Please set the `authToken` option. You can find information on how to generate a Sentry auth token here: https://docs.sentry.io/api/auth/" +
-          getTurborepoEnvPassthroughWarning("SENTRY_AUTH_TOKEN")
-      );
-    } else if (!options.org && !options.authToken.startsWith("sntrys_")) {
-      logger.warn(
-        "No org provided. Will not upload source maps. Please set the `org` option to your Sentry organization slug." +
-          getTurborepoEnvPassthroughWarning("SENTRY_ORG")
-      );
-    } else if (!options.project) {
-      logger.warn(
-        "No project provided. Will not upload source maps. Please set the `project` option to your Sentry project slug." +
-          getTurborepoEnvPassthroughWarning("SENTRY_PROJECT")
-      );
-    } else {
-      // This option is only strongly typed for the webpack plugin, where it is used. It has no effect on other plugins
-      const webpack_forceExitOnBuildComplete =
-        typeof options._experiments["forceExitOnBuildCompletion"] === "boolean"
-          ? options._experiments["forceExitOnBuildCompletion"]
-          : undefined;
+    // This option is only strongly typed for the webpack plugin, where it is used. It has no effect on other plugins
+    const webpack_forceExitOnBuildComplete =
+      typeof options._experiments["forceExitOnBuildCompletion"] === "boolean"
+        ? options._experiments["forceExitOnBuildCompletion"]
+        : undefined;
 
-      plugins.push(
-        debugIdUploadPlugin(
-          createDebugIdUploadFunction({
-            assets: options.sourcemaps?.assets,
-            ignore: options.sourcemaps?.ignore,
-            createDependencyOnSourcemapFiles,
-            dist: options.release.dist,
-            releaseName: options.release.name,
-            logger: logger,
-            handleRecoverableError: handleRecoverableError,
-            rewriteSourcesHook: options.sourcemaps?.rewriteSources,
-            sentryScope,
-            sentryClient,
-            sentryCliOptions: {
-              authToken: options.authToken,
-              org: options.org,
-              project: options.project,
-              silent: options.silent,
-              url: options.url,
-              vcsRemote: options.release.vcsRemote,
-              headers: options.headers,
-            },
-          }),
-          logger,
-          webpack_forceExitOnBuildComplete
-        )
-      );
-    }
+    plugins.push(
+      debugIdUploadPlugin(
+        createDebugIdUploadFunction({
+          sentryBuildPluginManager,
+        }),
+        logger,
+        sentryBuildPluginManager.createDependencyOnBuildArtifacts,
+        webpack_forceExitOnBuildComplete
+      )
+    );
 
     if (options.reactComponentAnnotation) {
       if (!options.reactComponentAnnotation.enabled) {
@@ -472,12 +182,7 @@ export function sentryUnpluginFactory({
 
     plugins.push(
       fileDeletionPlugin({
-        waitUntilSourcemapFileDependenciesAreFreed,
-        filesToDeleteAfterUpload: options.sourcemaps?.filesToDeleteAfterUpload,
-        logger,
-        handleRecoverableError,
-        sentryScope,
-        sentryClient,
+        sentryBuildPluginManager,
       })
     );
 
@@ -651,36 +356,45 @@ export function createRollupModuleMetadataInjectionHooks(injectionCode: string) 
 }
 
 export function createRollupDebugIdUploadHooks(
-  upload: (buildArtifacts: string[]) => Promise<void>
+  upload: (buildArtifacts: string[]) => Promise<void>,
+  _logger: Logger,
+  createDependencyOnBuildArtifacts: () => () => void
 ) {
+  const freeGlobalDependencyOnDebugIdSourcemapArtifacts = createDependencyOnBuildArtifacts();
   return {
     async writeBundle(
       outputOptions: { dir?: string; file?: string },
       bundle: { [fileName: string]: unknown }
     ) {
-      if (outputOptions.dir) {
-        const outputDir = outputOptions.dir;
-        const buildArtifacts = await glob(
-          [
-            "/**/*.js",
-            "/**/*.mjs",
-            "/**/*.cjs",
-            "/**/*.js.map",
-            "/**/*.mjs.map",
-            "/**/*.cjs.map",
-          ].map((q) => `${q}?(\\?*)?(#*)`), // We want to allow query and hashes strings at the end of files
-          {
-            root: outputDir,
-            absolute: true,
-            nodir: true,
-          }
-        );
-        await upload(buildArtifacts);
-      } else if (outputOptions.file) {
-        await upload([outputOptions.file]);
-      } else {
-        const buildArtifacts = Object.keys(bundle).map((asset) => path.join(path.resolve(), asset));
-        await upload(buildArtifacts);
+      try {
+        if (outputOptions.dir) {
+          const outputDir = outputOptions.dir;
+          const buildArtifacts = await glob(
+            [
+              "/**/*.js",
+              "/**/*.mjs",
+              "/**/*.cjs",
+              "/**/*.js.map",
+              "/**/*.mjs.map",
+              "/**/*.cjs.map",
+            ].map((q) => `${q}?(\\?*)?(#*)`), // We want to allow query and hashes strings at the end of files
+            {
+              root: outputDir,
+              absolute: true,
+              nodir: true,
+            }
+          );
+          await upload(buildArtifacts);
+        } else if (outputOptions.file) {
+          await upload([outputOptions.file]);
+        } else {
+          const buildArtifacts = Object.keys(bundle).map((asset) =>
+            path.join(path.resolve(), asset)
+          );
+          await upload(buildArtifacts);
+        }
+      } finally {
+        freeGlobalDependencyOnDebugIdSourcemapArtifacts();
       }
     },
   };
@@ -744,7 +458,6 @@ export function getDebugIdSnippet(debugId: string): string {
   return `;{try{let e="undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof globalThis?globalThis:"undefined"!=typeof self?self:{},n=(new e.Error).stack;n&&(e._sentryDebugIds=e._sentryDebugIds||{},e._sentryDebugIds[n]="${debugId}",e._sentryDebugIdIdentifier="sentry-dbid-${debugId}")}catch(e){}};`;
 }
 
-export { stringToUUID, replaceBooleanFlagsInCode } from "./utils";
-
-export type { Options, SentrySDKBuildFlags } from "./types";
 export type { Logger } from "./sentry/logger";
+export type { Options, SentrySDKBuildFlags } from "./types";
+export { replaceBooleanFlagsInCode, stringToUUID } from "./utils";
