+++ packages/bundler-plugin-core/src/plugins/sourcemap-deletion.ts
@@ -1,69 +1,17 @@
-import { glob } from "glob";
 import { UnpluginOptions } from "unplugin";
-import { Logger } from "../sentry/logger";
-import { safeFlushTelemetry } from "../sentry/telemetry";
-import fs from "fs";
-import { Scope } from "@sentry/core";
-import { Client } from "@sentry/types";
-import { HandleRecoverableErrorFn } from "../types";
+import { SentryBuildPluginManager } from "../api-primitives";
 
 interface FileDeletionPlugin {
-  handleRecoverableError: HandleRecoverableErrorFn;
-  waitUntilSourcemapFileDependenciesAreFreed: () => Promise<void>;
-  sentryScope: Scope;
-  sentryClient: Client;
-  filesToDeleteAfterUpload: string | string[] | Promise<string | string[] | undefined> | undefined;
-  logger: Logger;
+  sentryBuildPluginManager: SentryBuildPluginManager;
 }
 
 export function fileDeletionPlugin({
-  handleRecoverableError,
-  sentryScope,
-  sentryClient,
-  filesToDeleteAfterUpload,
-  waitUntilSourcemapFileDependenciesAreFreed,
-  logger,
+  sentryBuildPluginManager,
 }: FileDeletionPlugin): UnpluginOptions {
   return {
     name: "sentry-file-deletion-plugin",
     async writeBundle() {
-      try {
-        const filesToDelete = await filesToDeleteAfterUpload;
-        if (filesToDelete !== undefined) {
-          const filePathsToDelete = await glob(filesToDelete, {
-            absolute: true,
-            nodir: true,
-          });
-
-          logger.debug(
-            "Waiting for dependencies on generated files to be freed before deleting..."
-          );
-
-          await waitUntilSourcemapFileDependenciesAreFreed();
-
-          filePathsToDelete.forEach((filePathToDelete) => {
-            logger.debug(`Deleting asset after upload: ${filePathToDelete}`);
-          });
-
-          await Promise.all(
-            filePathsToDelete.map((filePathToDelete) =>
-              fs.promises.rm(filePathToDelete, { force: true }).catch((e) => {
-                // This is allowed to fail - we just don't do anything
-                logger.debug(
-                  `An error occurred while attempting to delete asset: ${filePathToDelete}`,
-                  e
-                );
-              })
-            )
-          );
-        }
-      } catch (e) {
-        sentryScope.captureException('Error in "sentry-file-deletion-plugin" buildEnd hook');
-        await safeFlushTelemetry(sentryClient);
-        // We throw by default if we get here b/c not being able to delete
-        // source maps could leak them to production
-        handleRecoverableError(e, true);
-      }
+      await sentryBuildPluginManager.deleteArtifacts();
     },
   };
 }
