+++ packages/esbuild-plugin/src/index.ts
@@ -222,16 +222,23 @@ function esbuildModuleMetadataInjectionPlugin(injectionCode: string): UnpluginOp
 }
 
 function esbuildDebugIdUploadPlugin(
-  upload: (buildArtifacts: string[]) => Promise<void>
+  upload: (buildArtifacts: string[]) => Promise<void>,
+  _logger: Logger,
+  createDependencyOnBuildArtifacts: () => () => void
 ): UnpluginOptions {
+  const freeGlobalDependencyOnDebugIdSourcemapArtifacts = createDependencyOnBuildArtifacts();
   return {
     name: "sentry-esbuild-debug-id-upload-plugin",
     esbuild: {
       setup({ initialOptions, onEnd }) {
         initialOptions.metafile = true;
         onEnd(async (result) => {
-          const buildArtifacts = result.metafile ? Object.keys(result.metafile.outputs) : [];
-          await upload(buildArtifacts);
+          try {
+            const buildArtifacts = result.metafile ? Object.keys(result.metafile.outputs) : [];
+            await upload(buildArtifacts);
+          } finally {
+            freeGlobalDependencyOnDebugIdSourcemapArtifacts();
+          }
         });
       },
     },
