+++ packages/webpack-plugin/src/index.ts
@@ -124,21 +124,28 @@ function webpackDebugIdInjectionPlugin(): UnpluginOptions {
 function webpackDebugIdUploadPlugin(
   upload: (buildArtifacts: string[]) => Promise<void>,
   logger: Logger,
+  createDependencyOnBuildArtifacts: () => () => void,
   forceExitOnBuildCompletion?: boolean
 ): UnpluginOptions {
   const pluginName = "sentry-webpack-debug-id-upload-plugin";
   return {
     name: pluginName,
     webpack(compiler) {
+      const freeGlobalDependencyOnDebugIdSourcemapArtifacts = createDependencyOnBuildArtifacts();
+
       compiler.hooks.afterEmit.tapAsync(pluginName, (compilation, callback: () => void) => {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
         const outputPath = (compilation.outputOptions.path as string | undefined) ?? path.resolve();
         const buildArtifacts = Object.keys(compilation.assets as Record<string, unknown>).map(
           (asset) => path.join(outputPath, asset)
         );
-        void upload(buildArtifacts).then(() => {
-          callback();
-        });
+        void upload(buildArtifacts)
+          .then(() => {
+            callback();
+          })
+          .finally(() => {
+            freeGlobalDependencyOnDebugIdSourcemapArtifacts();
+          });
       });
 
       if (forceExitOnBuildCompletion && compiler.options.mode === "production") {
